# group save_datasets result by file