Group the results of ``Scene.save_datasets(..., compute=False)`` by output file, when multiple files are to be written and one or more files have multiple RIODataset or RIOTag objects.  This is helpful when we want a wrapper per file and therefore a single ``da.store`` call for each file, which is in turn needed for https://github.com/pytroll/trollflow2/pull/168.

<!-- Describe what your PR does, and why -->
<!-- For works in progress choose "Create draft pull request" from the drop-down green button. -->

 - [ ] Closes #xxxx <!-- remove if there is no corresponding issue, which should only be the case for minor changes -->
 - [x] Tests added <!-- for all bug fixes or enhancements -->
 - [x] Fully documented <!-- remove if this change should not be visible to users, e.g., if it is an internal clean-up, or if this is part of a larger project that will be documented later -->