@djhoese 2022-11-16
This is an interesting idea. So in your trollflow callback PR you would do the grouping then pass each group to the callback? Biggest potential issue I see with that is if you have a dask array going to more than one file and you compute the groups separately, you'll likely end up computing each input image dask array multiple times. If you pass them all to `da.store` at the same time then this shouldn't matter. Or if you pass them to `da.store` separately but compute the Delayed results at the same time then that should be fine too.