@gerritholl 2022-11-16
The ones that go to the same file go to `da.store` at the same time.  I still send everything to `da.compute` at the same time.  The difference would be:

Now:

```python
obj = da.store(sources_for_all_files, targets_for_all_files, compute=False)
da.compute(obj)
```

Later, optionally:

```python
obj1 = da.store(sources_for_file_1, targets_for_file_1, compute=False)
obj2 = da.store(sources_for_file_2, targets_for_file_2, compute=False)
obj3 = da.store(sources_for_file_3, targets_for_file_3, compute=False)
da.compute([obj1, obj2, obj3])
```

I don't know what that would mean for performance, but I need approach 2 so that I can make a wrapper that does something as soon as an individual file is completed.  At least I can't think of another way to achieve that.