@gerritholl 2022-11-17
pre-commit did not complain when I did a commit locally...