@gerritholl 2022-11-17
I did some performance tests to compare between computing `da.store(src, targ, compute=False)` or `[da.store(s, t, compute=False) for (s, t) in zip(src, targ)]`, because in the context of the linked trollflow2 issue I find a severe impact on performance.  However, the performance penalty of having more `da.store` delayed objects to be computed is small.

```python
import os
import dask.array as da
import dask
from dask.diagnostics import Profiler, ResourceProfiler, visualize
from sattools.io import plotdir
from glob import glob
from satpy import Scene
from dask import delayed

seviri_files = glob("/media/nas/x21308/scratch/SEVIRI/202103300900/H-000*")
sc = Scene(filenames={"seviri_l1b_hrit": seviri_files})
names = sc.available_dataset_names()
sc.load(names)

(src, targ) = sc.save_datasets(
    writer="geotiff",
    filename=os.fspath(plotdir() / "test-{name}.tif"),
    fill_value=0,
    compute=False)
#delayeds = da.store(src, targ, compute=False)
delayeds = [da.store(s, t, compute=False) for (s, t) in zip(src, targ)]
with Profiler() as prof, ResourceProfiler(dt=0.05) as rprof:
    dask.compute(delayeds)
visualize([prof, rprof], show=False, save=True, filename=os.fspath(plotdir() /
    f"dask-profile-many-store.html"))
```

Using just `da.store(src, targ, compute=False)`:

![image](https://user-images.githubusercontent.com/500246/202520914-9577d943-9bdf-4988-ad59-7cd3cdc298d4.png)

Using `[da.store(s, t, compute=False) for (s, t) in zip(src, targ)]`, RAM increases from 1.38 GB to 1.45 GB, with wall clock time increasing from 0:13.21 to 0:13.53.

![image](https://user-images.githubusercontent.com/500246/202521347-380f75c7-98ca-468b-a75c-3e6909cc6266.png)