@djhoese 2022-11-17
A couple things come to mind:

1. I think you should find a larger test case. One that takes at least 30 seconds if not closer to multiple minutes. One that only takes a few seconds feels like it would be heavily influenced by disk/OS caches.
2. There was (or still exists) a bug in dask.array.store where some cases aren't as optimized as they should be. I believe I filed a bug with dask or contributed to an existing one, but I don't remember if it was ever resolved. It may be one where I was expected to fix it but I didn't understand the problem enough to fix it.