@gerritholl 2022-11-18
Loading FCI with 4 composites and 3 channels, then resampling to an equirectangular 2km-grid containing the full FCI field of view.  This reveals an increase in RAM by 19%, but no increase in runtime.  The test machine has 8 CPUs and 128 GB RAM.  There is no additional increase of resources caused by adding a wrapper function.

With a single `da.store(..., compute=False)` call passed to `dask.compute`:

3:16.47, 33.6 GB

![image](https://user-images.githubusercontent.com/500246/202676702-3998c719-dc68-4905-89c6-61fa8605a6ea.png)

[dask-profile-no-wrap-single-store.html.gz](https://github.com/pytroll/satpy/files/10040323/dask-profile-no-wrap-single-store.html.gz)

With multiple `da.store(..., compute=False)` calls passed to `dask.compute`, no wrapper:

3:18.61, 40.1 GB

[dask-profile-no-wrap-multi-store.html.gz](https://github.com/pytroll/satpy/files/10040321/dask-profile-no-wrap-multi-store.html.gz)

![image](https://user-images.githubusercontent.com/500246/202676820-3c984f9b-7384-48e2-91bd-c9b9ca6674f3.png)

With multiple `da.store(..., compute=False)` calls passed to `dask.compute`, and a noop wrapper:

3:17.38, 39.8 GB

![image](https://user-images.githubusercontent.com/500246/202677115-a21bcf7a-ccae-45d9-a47f-1d39d4eb1a09.png)

[dask-profile-wrap-multi-store.html.gz](https://github.com/pytroll/satpy/files/10040324/dask-profile-wrap-multi-store.html.gz)

Timings performed with:

```python
import hdf5plugin
import os
import dask.array as da
import dask
from dask.diagnostics import Profiler, ResourceProfiler, visualize
from glob import glob
from satpy import Scene
from dask import delayed

fci_files = glob("/media/x21308/MTG_test_data/2022_05_MTG_Testdata/RC0042/*BODY*.nc")
sc = Scene(filenames={"fci_l1c_nc": fci_files})
names = ["natural_color", "true_color", "airmass", "cimss_cloud_type",
        "ir_105", "vis_04", "vis_06"]
sc.load(names)

def noop(obj):
    """Do nothing."""
    return obj

ls = sc.resample("nq0002km")
(src, targ) = ls.save_datasets(
    writer="geotiff",
    filename="/media/x21308/scratch/dask-tests/{start_time:%Y%m%d%H%M}-{platform_name}-{sensor}-{area.area_id}-{name}-dask-test.tif",
    fill_value=0,
    compute=False)
#delayeds = da.store(src, targ, compute=False)
delayeds = [da.store(s, t, compute=False) for (s, t) in zip(src, targ)]
#delayeds = [delayed(noop)(da.store(s, t, compute=False)) for (s, t) in zip(src, targ)]
with Profiler() as prof, ResourceProfiler(dt=0.05) as rprof:
    dask.compute(delayeds)
visualize([prof, rprof], show=False, save=True, filename="/tmp/dask-profile-no-wrap-multi-store.html")
```

and then called with

```bash
\time -v python dask-delay-wrapper-resource-problem-large.py
```