@gerritholl 2022-11-18
When I increase this to 7 composites and 6 channels:

- single `da.store` call: 4:11.41, 38.1 GB
- multiple `da.store` calls: 4:21.06, 43.2 GB

so the differences are relatively smaller