@gerritholl 2022-11-18
When storing 7 composites and 6 channels to `ninjogeotiff`, which has some tags that are also dask-based (`RIOTag`), we get:

* single call: 3:26.81, 35.2 GB
* one call per target: 5:30.63, 37.4 GB
* one call per output file (using the splitter): 3:36.22, 38.7 GB

I'm getting inconsistent results and remain confused.