@gerritholl 2022-11-18
If I wrap the `da.store(..., compute=False)` calls with a `with dask.config.set({"optimization.fuse.active": False})`, such as suggested by @djhoese in https://github.com/dask/dask/issues/8380#issuecomment-969248814, RAM goes up even more.  The one call per target is faster at 4:08.82, but uses 47.6 GB RAM.

But maybe that's not at all surprising, considering I have no idea what I'm doing ;)