@gerritholl 2022-11-23
Comparison:

| Enhancing | Resampling | DWD channels | Composites | scheduler | time baseline | RAM baseline | time callback | RAM callback 
| --- | --- | --- | --- | --- | --- | --- | --- | --- |
| True | eqc 1km/2km | True | 7 | threads | 08:56.4 | 83.7 GB | 11:32.9 | 64.6 GB |
| True | eqc 3km | True | 3 | threads | 01:48.5 | 15.4 GB | 02:00.0 | 15.4 GB |
| True | False | True | 3 | threads | 02:04.4 | 6.51 GB | 01:53.9 | 8.21 GB |
| True | eqc 3km | False | 3 | threads | 01:26.6 | 15.4 GB | 01:39.2 | 15.4 GB |
| True | False | False | 3 | threads | 01:26.3 | 8.74 GB | 01:16.1 | 7.67 GB |
| True | False | False | False | threads | 01:10.4 | 6.19 GB | 01:07.3 | 5.20 GB |
| False | False | False | False | threads | 01:25.0 | 6.47 GB | 01:17.2 | 6.33 GB |
| False | False | False | False | synchronous | 01:52.0 | 5.88 GB | 01:25.8 | 5.43 GB |

This reads full disc FCI test data and writes at least all channels in various configurations.

* Callback: a single callback that closes the targets, nothing else
* DWD channels: if true, replace single channels (with default stretching) by single-channel composites with enhancements with crude stretching and inverting IR imagery
* Composites: in addition to all FCI channels, also do airmass, dust, and ash RGBs

The first line represents the experimental processing with all products.

So, it seems that adding the callback makes things slower only when we're resampling or when RAM usage is high.  Without resampling, adding the callback may even make things faster.